loan_manager/requirements.txt
PyQt6>=6.6.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0
requests>=2.31.0
PyJWT>=2.8.0
bcrypt>=4.1.0
openpyxl>=3.1.2
pyinstaller>=6.3.0
python-multipart>=0.0.6
wechatpayv3>=0.10.0
python-alipay-sdk>=3.0.0
